+++ changelog.md
@@ -5,6 +5,8 @@
 > [!IMPORTANT]
 > This version is not released yet and is under active development.
 
+- Add support for auto-conversion of GitHub alerts into MyST admonitions in Sphinx extension.
+
 ## [7.1.0 (2025-11-21)](https://github.com/kdeldycke/click-extra/compare/v7.0.1...v7.1.0)
 
 - Add support for aliases in `EnumChoice` type.